requirements.txt
appnope==0.1.0
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.3
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cv==1.0.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.36.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==3.1.4
django-webpack-loader==0.6.0
djangorestframework==3.11.0
entrypoints==0.3
fastai==1.0.60
fastprogress==0.2.2
gunicorn==19.7.1
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
munch==2.5.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
plac==1.1.3
preshed==3.0.2
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2==2.8.4
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.6.1
pylint==2.4.4
pynvx==1.0.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
spacy==2.2.4
sqlparse==0.3.0
srsly==1.0.2
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.44.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==4.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
